Add dedicated filter service mock type
